+++ src/main/views/entities/interfaces/Animated.java
@@ -7,14 +7,10 @@ import main.utils.SpriteAnimation;
 public interface Animated {
   default void animate(ImageView imageView) {
     /* count and columns params are not modeled that's why are passed as magic numbers */
-    final double count = 2;
-    final double columns = 2;
+    final float count = 2;
+    final float columns = 2;
     
-    final Animation animation = new SpriteAnimation(
-          imageView,
-          count,
-          columns
-    );
+    final Animation animation = new SpriteAnimation(imageView, count, columns);
     
     animation.setCycleCount(Animation.INDEFINITE);
     animation.play();
